now it works with double precision floating point numbers (SSE2 instructions) + fixed bug in abs
--HG--
branch : doublepresso

+++ MathPresso/MathPresso_AST.cpp
@@ -255,7 +255,7 @@ mreal_t ASTOperator::evaluate(void* data) const
     {
       mreal_t vl = _left->evaluate(data);
       mreal_t vr = _right->evaluate(data);
-      result = fmodf(vl, vr);
+      result = fmod(vl, vr);
       break;
     }
   }
